backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
anthropic==0.28.0
python-multipart==0.0.9
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.129.0
httpx==0.27.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyPDF2==3.0.1
python-docx==1.1.2
